% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{outpack_copy_files}
\alias{outpack_copy_files}
\title{Copy files from a packet}
\usage{
outpack_copy_files(id, files, dest, root = NULL)
}
\arguments{
\item{id}{Optionally, an outpack id via \link{outpack_id}. If
not given a new id will be generated.}

\item{files}{A named character vector of files; the name
corresponds to the name within the current packet, while the
value corresponds to the name within the upstream packet}

\item{dest}{The directory to copy into}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing, invisibly. Primarily called for its side effect
of copying files from a packet into the directory \code{dest}
}
\description{
Copy files from a packet to anywhere. Similar to
\link{outpack_packet_use_dependency} except that this is not
used in an active packet context. You can use this function to
pull files from an outpack root to a directory outside of the
control of outpack, for example.
}
