% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{outpack_init}
\alias{outpack_init}
\title{Create outpack root}
\usage{
outpack_init(root, path_archive = "archive", use_file_store = FALSE)
}
\arguments{
\item{root}{Path to use.  This path may exist, but it is an error
to call this on a path that has already been initialised.}

\item{path_archive}{Path to the archive directory, used to store
human-readable copies of packets.  If \code{NULL}, no such copy is
made, and \code{file_store} must be \code{TRUE}}

\item{use_file_store}{Logical, indicating if we should use a
content-addressable file-store as the source of truth for
packets.  If \code{archive} is non-\code{NULL}, the file-store will be
used as the source of truth and the duplicated files in archive
exist only for convenience.}
}
\value{
Invisibly, an \code{outpack_root} object; these will be
}
\description{
Create a new outpack root.
}
