% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{outpack_location_add}
\alias{outpack_location_add}
\title{Add a new location}
\usage{
outpack_location_add(name, path, root = NULL)
}
\arguments{
\item{name}{The short name of the location to use.  Cannot be in
use, and cannot be one of \code{local} or \code{orphan}}

\item{path}{The path to the directory containing another outpack
archive.}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Add a new location.
}
\section{Warning}{


The API here will change as we move to support different types of
locations.
}

