% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{outpack_location_list}
\alias{outpack_location_list}
\title{List known pack locations}
\usage{
outpack_location_list(root = NULL)
}
\arguments{
\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
A character vector of location names. The special name
\code{local} will always be present.
}
\description{
List known locations.
}
