% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{outpack_location_pull_packet}
\alias{outpack_location_pull_packet}
\title{Pull a single packet from a location}
\usage{
outpack_location_pull_packet(id, location, root = NULL)
}
\arguments{
\item{id}{The id of the packet to pull}

\item{location}{The name of the location to pull from.  Later we
will relax this (see mrc-3030)}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Pull a packet (all files) from a location into this archive. This
will make files available for use as dependencies (e.g., with
\link{outpack_packet_use_dependency})
}
\details{
The behaviour of this function will vary depending on whether or
not the destination outpack repository (i.e., \code{root}) uses a file
store or not.  If it does, then we simply import the unknown files
into the store, and this will always be fairly efficient.  If no
file store is used then for the time being we pull all files from
the upstream location, even if this means copying a file we
already know about elsewhere in the outpack archive.  We will
improve this in a future version.
}
