% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{outpack_location_pull_tree}
\alias{outpack_location_pull_tree}
\title{Pull tree of packets}
\usage{
outpack_location_pull_tree(id, location, root = NULL)
}
\arguments{
\item{id}{The id of the packet to pull}

\item{location}{The name of the location to pull from.  Later we
will relax this (see mrc-3030)}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Nothing
}
\description{
Pull a consistent tree of packets from a location. Unlike
\link{outpack_location_pull_packet}, this pulls all packets
that were used in creating the target packet, and pulling a packet
that is already unpacked is not an error.
}
