% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{outpack_location_push}
\alias{outpack_location_push}
\title{Push packets into a location}
\usage{
outpack_location_push(ids, location, root)
}
\arguments{
\item{ids}{A vector of packets}

\item{location}{The name of the location to push to}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}
}
\value{
Undecided
}
\description{
Push packets into a location. Potentially a costly operation.
}
