% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{outpack_metadata_create}
\alias{outpack_metadata_create}
\title{Create outpack metadata}
\usage{
outpack_metadata_create(
  path,
  name,
  id,
  time,
  inputs,
  outputs,
  depends = NULL,
  parameters = NULL,
  session = NULL,
  extra = NULL,
  hash_algorithm = "sha256"
)
}
\arguments{
\item{path}{Directory holding the packet}

\item{name}{The name of the packet}

\item{id}{The unique identifier}

\item{time}{A list of times. Must contain the elements \code{start} and \code{end} (as \code{POSIXt} objects) and may contain any}

\item{inputs}{A character vector of inputs}

\item{outputs}{A character vector of outputs}

\item{depends}{Optionally, information about dependencies on other
packets.  This must a list of objects created by
\code{outpack::outpack_metadata_depends}, and all files referenced
should be found within \code{inputs}}

\item{parameters}{Optionally, a list of named key/value
parameters.  Each key must be string and each value may be a
boolean, number or string (no vectors are allowed).}

\item{session}{Optionally, information about the running session.
If omitted, then default session information is included.}

\item{extra}{Optionally, arbitrary additional metadata as needed.
This will be passed through \code{jsonlite::toJSON} with arguments
\verb{pretty = FALSE, auto_unbox = FALSE, json_verbatim = TRUE, na = "null", null = "null"}, so plan accordingly.}

\item{hash_algorithm}{Optionally, the hash algorithm to use.  The
default is \code{sha256} which is what git uses and should be good
for most uses.}
}
\value{
A \code{json} string
}
\description{
Create outpack metadata.
}
