% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{outpack_metadata_load}
\alias{outpack_metadata_load}
\title{Load outpack metadata}
\usage{
outpack_metadata_load(json)
}
\arguments{
\item{json}{A path to generated json, or the json itself as a
string.}
}
\value{
A list with the deserialised metadata
}
\description{
Load outpack metadata as an R object.  Note that due to the
difficulties of serialising R objects to JSON this cannot be
directly re-converted to JSON (but as the outpack metadata is
immutable, that should not be a big limitation.
}
