% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{outpack_metadata_validate}
\alias{outpack_metadata_validate}
\title{Validate metadata against schema}
\usage{
outpack_metadata_validate(json)
}
\arguments{
\item{json}{A path to generated json, or the json itself as a
string.}
}
\value{
Logical \code{TRUE} (invisibly) if valid, error otherwise.
}
\description{
Validate metadata against the schema
}
