% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet.R
\name{outpack_packet_start}
\alias{outpack_packet_start}
\alias{outpack_packet_cancel}
\alias{outpack_packet_current}
\alias{outpack_packet_end}
\alias{outpack_packet_run}
\alias{outpack_packet_use_dependency}
\title{Start, interact with, and end a packet build}
\usage{
outpack_packet_start(path, name, parameters = NULL, root = NULL)

outpack_packet_cancel()

outpack_packet_current()

outpack_packet_end()

outpack_packet_run(script, envir = .GlobalEnv)

outpack_packet_use_dependency(id, files)
}
\arguments{
\item{path}{Path to the build / output directory.}

\item{name}{The name of the packet}

\item{parameters}{Optionally, a named list of parameters.  The
names must be unique, and the values must all be non-NA scalar
atomics (logical, integer, numeric, character)}

\item{root}{The outpack root. Will be searched for from the
current directory if not given.}

\item{script}{Path to the script within the packet directory (a
relative path).  This function can be safely called multiple
times within a single packet run (or zero times!) as needed.}

\item{envir}{Environment in which to run the script}

\item{id}{The id of an existing packet to use files from}

\item{files}{A named character vector of files; the name
corresponds to the name within the current packet, while the
value corresponds to the name within the upstream packet}
}
\value{
Invisibly, a copy of the packet data
}
\description{
Start a packet build (\code{outpack_packet_start}), end one
(\code{outpack_packet_cancel}, \code{outpack_packet_end}) and interact with
the current packet (\code{outpack_packet_use_dependency},
\code{outpack_packet_run})
}
