% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet.R
\name{outpack_packet_add_git}
\alias{outpack_packet_add_git}
\title{Add git metadata to current packet}
\usage{
outpack_packet_add_git()
}
\value{
Invisibly, a logical indicating if git information was
successfully added. This will be \code{FALSE} if the repository is
not under version control
}
\description{
Add git metadata to the current packet. This stores some key
metadata (current branch, commit, and urls of known remote
repositories) to the packet metadata.
}
