% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet.R
\name{outpack_packet_file_mark}
\alias{outpack_packet_file_mark}
\alias{outpack_packet_file_list}
\title{Mark files during packet run}
\usage{
outpack_packet_file_mark(files, status = "immutable")

outpack_packet_file_list()
}
\arguments{
\item{files}{A character vector of relative paths}

\item{status}{A status to mark the file with. Must be "immutable"
for now, later we will expand this.}
}
\value{
\itemize{
\item \code{outpack_packet_file_mark} returns nothing
\item \code{outpack_packet_file_list} returns a \link{data.frame} with columns
\code{path} and \code{status} (\code{immutable}, \code{ignored}, \code{unknown})
}
}
\description{
Mark file within an in-progress packet. This will store the hash
of the file within the internal outpack structures and force an
error if the file is changed or deleted later.  The function
\code{\link[=outpack_packet_list_unmarked]{outpack_packet_list_unmarked()}} will report on unmarked
files present in the directory.
}
