% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packet.R
\name{outpack_packet_mark_file}
\alias{outpack_packet_mark_file}
\alias{outpack_packet_list_unmarked}
\title{Mark files during packet run}
\usage{
outpack_packet_mark_file(files)

outpack_packet_list_unmarked()
}
\arguments{
\item{files}{A character vector of relative paths}
}
\value{
Nothing
}
\description{
Mark file within an in-progress packet. This will store the hash
of the file within the internal outpack structures and force an
error if the file is changed or deleted later.  The function
\code{\link[=outpack_packet_list_unmarked]{outpack_packet_list_unmarked()}} will report on unmarked
files present in the directory.
}
