% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{outpack_session_info}
\alias{outpack_session_info}
\title{Create session information}
\usage{
outpack_session_info(info)
}
\arguments{
\item{info}{Output from \code{\link[=sessionInfo]{sessionInfo()}}}
}
\value{
A list with information on the platform and packages
}
\description{
Create a reasonable \code{session} entry using output from \code{\link[=sessionInfo]{sessionInfo()}}
}
\examples{
dat <- outpack::outpack_session_info(sessionInfo())
jsonlite::toJSON(dat, pretty = TRUE)
}
