% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_index.R
\name{query_index}
\alias{query_index}
\title{Query index}
\description{
Class for managing the active index whilst evaluating a query
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{index}}{The packet index}

\item{\code{depends}}{Named list of data frames. Names are packet ids, values
are packets dependend on by this packet id (i.e. its parents).}

\item{\code{root}}{The outpack root object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-query_index-new}{\code{query_index$new()}}
\item \href{#method-query_index-scope}{\code{query_index$scope()}}
\item \href{#method-query_index-get_packet_depends}{\code{query_index$get_packet_depends()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_index-new"></a>}}
\if{latex}{\out{\hypertarget{method-query_index-new}{}}}
\subsection{Method \code{new()}}{
Create a new query_index object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{query_index$new(root, index, depends)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{root}}{The outpack root object}

\item{\code{index}}{The packet index as a data frame}

\item{\code{depends}}{Named list of data frames. Names are packet ids, values
are packets dependend on by this packet id (i.e. its parents).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_index-scope"></a>}}
\if{latex}{\out{\hypertarget{method-query_index-scope}{}}}
\subsection{Method \code{scope()}}{
Scope the index. This will filter the active index on the scope
and save the scoped index for use later
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{query_index$scope(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{The ids to scope the index on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, called for side effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_index-get_packet_depends"></a>}}
\if{latex}{\out{\hypertarget{method-query_index-get_packet_depends}{}}}
\subsection{Method \code{get_packet_depends()}}{
Get the ids of packets which this packet depends to a specified level
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{query_index$get_packet_depends(id, depth)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the packet to get parents of}

\item{\code{depth}}{Depth of parents to get, \code{depth} 1 gets immediate parents
\code{depth} 2 gets parents and parents of parents, \code{depth} Inf will
recurse the whole tree to get all parents}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The ids of the parents of this packet
}
}
}
