% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_deparse.R
\name{outpack_query_format}
\alias{outpack_query_format}
\title{Format outpack query for displaying to users}
\usage{
outpack_query_format(query)
}
\arguments{
\item{query}{The outpack query to print}
}
\value{
Query expression as a string
}
\description{
Format outpack query for displaying to users
}
\examples{
outpack_query_format(quote(name == "example"))
}
